Adding an override for focusFollowsMouse upon entering a window.
There already was a complementary override upon leaving a window.

Since I'm using this override for Firefox, the risk of accidentally
closing windows not being pointed at is not insigificant. For that
reason, a replacement binding for Mod+Shify+c which closes the pointed
window rather than the focused one was introduced.