Turning off the cedesFocusEventHook override for now.
Cf. a90fe843f61343e722c354d2cf7c05700681d060 , which introduced the
override. After using it for a couple days, my impression is that the
slight convenience it adds doesn't trump the annoyance of inconsistent
focusing behaviours. Later, I'll investigate the possibility of having
a stateful toggle for this kind of override.